requirements.txt
fastapi 
uvicorn
python-multipart
openai 
python-dotenv
gtts 
pydantic 
groq
PyPDF2
python-docx
pydantic-settings
pdfplumber
pyttsx3
